@@ -1,14 +1,19 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.21'
+   // ext.kotlin_version = '1.3.21'
+    ext.kotlin_version = "1.4.32"
     repositories {
         google()
+        mavenCentral()
         jcenter()
+
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.2'
+        //classpath 'com.android.tools.build:gradle:3.4.3'
+      //  classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "com.android.tools.build:gradle:7.1.3"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -18,8 +23,9 @@ buildscript {
 allprojects {
     repositories {
         google()
+        mavenCentral()
         jcenter()
-        
+        maven { url "https://maven.google.com/"}
     }
 }
 
