@@ -2,7 +2,7 @@ package com.utility.finmartcontact.CallLog
 
 //import kotlinx.android.synthetic.main.layout.call_log_item.xml.view.*
 import android.content.Context
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
