@@ -2,9 +2,9 @@ package com.utility.finmartcontact.utility
 
 import android.app.Activity
 import android.content.pm.PackageManager
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AlertDialog
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import androidx.appcompat.app.AlertDialog
 import android.widget.Toast
 
 class RequestPermission(val activity: Activity, val list: List<String>, val code: Int) {
